Fix CI failures: formatting, typing, and linting
- Fixed unused `secrets` import in `pickaladder/group/routes.py`.
- Added null check for `email` in `pickaladder/group/routes.py` to resolve mypy errors.
- Added `# noqa: PLR0913` to `invite_by_email` in `pickaladder/group/services/group_service.py` to suppress too many arguments warning.
- Reformatted and sorted imports across modified files.
