+++ pickaladder/static/style.css
@@ -336,27 +336,3 @@ body {
         justify-content: space-between;
     }
 }
-/* Profiles & Avatars */
-.profile-picture {
-    width: 120px;
-    height: 120px;
-    object-fit: cover;
-}
-
-.profile-picture-large {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.profile-picture-thumbnail {
-    width: 40px;
-    height: 40px;
-    object-fit: cover;
-}
-
-.avatar {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
